Moved out triggers from URLBar
